Add support for [tags] in comments
A tag of [OTR] (case insensitive) means that we're *NOT* going to
publish the link.  It's off the record.

Currently HTML-encoding the tags has the effect of tagging posts with
e.g. the "&hearts;" tag;  not sure it's possible to have non-ASCII tags
via the Tumblr API right now.

On the other hand, not encoding them has the effect of triggering an API
error, so let's at least make the post succeed and eventual other tags
be present.